+++ lib/url.js
@@ -4,7 +4,7 @@ function importPunycode() {
   try {
     return process.binding('icu');
   } catch (e) {
-    return require('punycode');
+    return require('internal/punycode');
   }
 }
