+++ test/message/core_line_numbers.out
@@ -1,10 +1,10 @@
-punycode.js:42
+internal/punycode.js:42
 	throw new RangeError(errors[type]);
 	^
 
 RangeError: Invalid input
-    at error (punycode.js:42:*)
-    at Object.decode (punycode.js:*:*)
+    at error (internal/punycode.js:42:*)
+    at Object.decode (internal/punycode.js:*:*)
     at Object.<anonymous> (*test*message*core_line_numbers.js:*:*)
     at Module._compile (module.js:*:*)
     at Object.Module._extensions..js (module.js:*:*)
